i3290: Initial commit of client/common files
This set of sanitized conf files represents the configuration
used on our client nodes that forward, with only a few
exceptions, all log messages to a remote receiver.

As of this commit we have multiple nodes with large DA
queues that are not draining properly. A reboot seems to
correct whatever "state" issue is in play and the DA queue
quickly drains.

refs rsyslog/rsyslog#3290